src/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
black==24.4.2
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.6.0
Django==5.0.7
django-extensions==3.2.3
django-render-block==0.10
h11==0.14.0
Jinja2==3.1.4
kombu==5.4.0
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pyactiveresource==2.2.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.8
ruff==0.5.5
ShopifyAPI==12.6.0
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.3
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
